requirements.txt
bottle
bottle-cork